* 'develop' of github.com:genropy/genropy: (144 commits)
  pasteFromClipboard, xlsx integration
  dynamic load resource and genro plugin
  removed unused import
  new 'gnr dev bugreport' cli tool for reporting
  print timing on stderr, to avoid output capture mixup
  gnr.sql package provides AVAILABLE_DB_IMPLEMENTATIONS (static list of all adapters in the frameworks) and a hash ADAPTERS_BY_CAPABILITY, to query adapters which provides a specific capability.
  removed "test" in __main__ since it references an instance which doesn't exists in the framework itself, and use a non-existing method of the tests class.
  output dev notes in json format and Bag xml format
  Fix new_window
  db adapters can express dependencies on external tools to be invoke via subprocess() or pipes, and these are checked at startup.
  F401 linter cleanup
  F401 linter cleanup
  Re-introduce the dropdb invocation in migration tests Use same db name for psycopg2 tests Align connect behaviour between psycopg2 and psycopg3, raising exception on OperationalError Now the 2 adapters shares the same capabilities. o/
  removed wrong import introduced by an allucinated LLM
  linter import cleanup
  capabilities replaced with class methods (could be queried before creating an instance)
  fix keepselection in grouper
  handle missing package for a relation, providing specific error message.
  removed exception intercept in order to debug corretly relation issues
  Login for wsheadless clients
  ...

+++ projects/gnrcore/packages/sys/resources/services/storage/aws_s3.py
@@ -106,7 +106,7 @@ class Service(StorageService):
     def __init__(self, parent=None, bucket=None,
         base_path=None, aws_access_key_id=None,
         aws_secret_access_key=None, aws_session_token=None,
-        region_name=None, url_expiration=None, **kwargs):
+        region_name=None, url_expiration=None, write_in_local=None, readonly=None, **kwargs):
         self.parent = parent
         self.bucket = bucket
         self.base_path = (base_path or '').rstrip('/')
@@ -115,6 +115,11 @@ class Service(StorageService):
         self.aws_session_token=aws_session_token
         self.region_name=region_name
         self.url_expiration = url_expiration or 3600
+        has_sys = 'gnrcore:sys' in self.parent.gnrapp.config['packages']
+        secondary = has_sys and self.parent.gnrapp.config['packages'].getAttr('gnrcore:sys').get('secondary')
+        local = getattr(parent,'_local_mode', False)
+        local_readonly = (local or secondary) and not write_in_local
+        self.readonly = readonly or local_readonly
     @property
     def location_identifier(self):
         return 's3/%s/%s' % (self.region_name, self.bucket)
@@ -158,6 +163,8 @@ class Service(StorageService):
         pass
 
     def mkdir(self, *args, **kwargs):
+        if self.readonly:
+            return
         with self.open(*args+('.gnrdir',),mode='w') as dotfile:
             dotfile.write('.gnrdircontent')
 
@@ -203,6 +210,11 @@ class Service(StorageService):
         mode = kwargs.get('mode', 'r')
         keep = kwargs.get('keep', False)
         internalpath = self.internal_path(*args)
+        if self.readonly:
+            if 'b' in mode:
+                mode='rb'
+            else:
+                mode='r'
         return S3TemporaryFilename(bucket=self.bucket, key=internalpath,
             s3_session=self._session, mode=mode, keep=keep)
 
@@ -248,9 +260,13 @@ class Service(StorageService):
         return self._boto_client
 
     def delete_file(self, *args):
+        if self.readonly:
+            return
         self._client.delete_object(Bucket=self.bucket, Key=self.internal_path(*args))
 
     def delete_dir(self, *args):
+        if self.readonly:
+            return
         prefix = self.internal_path(*args)
         client = self._client
         prefix = prefix.strip('/')+'/' if prefix else ''
@@ -300,17 +316,26 @@ class Service(StorageService):
 
     def open(self, *args, **kwargs):
         kwargs['mode'] = kwargs.get('mode', 'rb')
+        if self.readonly:
+            if 'b' in kwargs['mode']:
+                kwargs['mode'] = 'rb'
+            else:
+                kwargs['mode'] = 'r'
         so_open.DEFAULT_BUFFER_SIZE = 1024 * 1024
         return so_open("s3://%s/%s"%(self.bucket,self.internal_path(*args)),
             transport_params={'session':self._session, 'client': self._client}, **kwargs)
 
 
     def duplicateNode(self, sourceNode=None, destNode=None): # will work only in the same bucket
+        if self.readonly:
+            return
         self._s3_copy(source_bucket=sourceNode.service.bucket,
             source_key=sourceNode.internal_path,
             dest_bucket=destNode.service.bucket, dest_key=destNode.internal_path)
 
     def renameNode(self, sourceNode=None, destNode=None):
+        if self.readonly:
+            return
         self._s3_copy(source_bucket=sourceNode.service.bucket,
             source_key=sourceNode.internal_path,
             dest_bucket=destNode.service.bucket, dest_key=destNode.internal_path)
@@ -362,6 +387,8 @@ class ServiceParameters(BaseComponent):
     def service_parameters(self,pane,datapath=None,**kwargs):
         bc = pane.borderContainer()
         fb = bc.contentPane(region='top').formbuilder(datapath=datapath)
+        fb.checkbox(value='^.readonly',lbl='Read Only')
+        fb.checkbox(value='^.write_in_local',lbl='Write enabled on local/secondary machine')
         fb.textbox(value='^.bucket',lbl='Bucket')
         fb.textbox(value='^.base_path',lbl='Base path')
         fb.textbox(value='^.aws_access_key_id',lbl='Aws Access Key Id')
